+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 > - 👓 Spec Compliance
 > - 🚀 New Feature
 > - 🐛 Bug Fix
+> - 👎 Deprecation
 > - 📝 Documentation
 > - 🏠 Internal
 > - 💅 Polish
@@ -31,6 +32,12 @@
 | `web-streams-polyfill/ponyfill` | `web-streams-polyfill/es5` | ES5+ ponyfill |
 | `web-streams-polyfill/ponyfill/es6` | `web-streams-polyfill` | ES2015+ ponyfill |
 | `web-streams-polyfill/ponyfill/es2018` | `web-streams-polyfill` | ES2015+ ponyfill |
+
+## v3.2.0 (2021-11-06)
+
+* 👎 Deprecate `WritableStreamDefaultController.abortReason` ([#102](https://github.com/MattiasBuelens/web-streams-polyfill/pull/102))
+  * Use `WritableStreamDefaultController.signal.reason` instead.
+* 👓 Align with [spec version `4b6b93c`](https://github.com/whatwg/streams/tree/4b6b93c69e531e2fe45a6ed4cb1484a7ba4eb8bb/) ([#103](https://github.com/MattiasBuelens/web-streams-polyfill/pull/103))
 
 ## v3.1.1 (2021-09-06)